Adapt mctools/G4 EM physics to Geant4 9.6 - 10.4
Use G4Version macros to adapt implementation to Geant4 interfaces from
9.6 to 10.4. Allows any version in this series to be used.

The major changes are the removal of certain models between 9.6 and
10.0. Where the models have been folded into one (e.g. G4UrbanMsc),
the old names are retained in the model factory, but the new object
is returned. Otherwise, replacements are per the recomendations in
Geant4 10.0's release notes.

As mctools/G4 is a generic application, clients updating to and
using a new Geant4 version should of course revalidate their
simulation outputs.