Update mctools/G4 neutron physics to Geant4 9.6-10.4
Use G4Version macros and interface updates to allow neutron
physics constructor to be compiled against any Geant4 between
9.6 to 10.4.

Due to the significant changes in hadronic models, the changes
are not feature complete. Compilation works in geant4 10.0 and better,
but the coherence of the processes and models cannot as yet. Clients
should prefer to use reference physics lists.